# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API contract for the STT Benchmark Command Center
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /benchmark/bulks:
    get:
      operationId: listBulks
      tags: [benchmark]
      summary: List evaluation bulks (FR-BLK-1), newest first
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/BulkStatus"
      responses:
        "200":
          description: Bulks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bulk"
    post:
      operationId: createBulk
      tags: [benchmark]
      summary: Create a bulk from selection criteria. Resolves and FREEZES the call set, estimates cost, and either launches immediately (status running) or parks at awaiting_confirmation when the estimate exceeds the server cost threshold (FR-BLK-5) -- re-POST with confirm:true or call /launch to proceed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkInput"
      responses:
        "201":
          description: Created bulk (running, or awaiting_confirmation when over the cost threshold)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bulk"
        "400":
          description: Invalid criteria or nothing matched
        "409":
          description: Bulk name already in use (names are unique, FR-BLK-2)
  /benchmark/bulks/{bulkId}:
    get:
      operationId: getBulk
      tags: [benchmark]
      summary: Bulk detail with live progress aggregates (FR-EXC-4, FR-BLK-13)
      parameters:
        - name: bulkId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bulk with progress counters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDetail"
        "404":
          description: Bulk not found
  /benchmark/bulks/{bulkId}/launch:
    post:
      operationId: launchBulk
      tags: [benchmark]
      summary: Launch a draft/awaiting_confirmation bulk -- fans the frozen call set into shard runs (FR-BLK-3). Async; poll GET /benchmark/bulks/{bulkId}.
      parameters:
        - name: bulkId
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Launch started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bulk"
        "404":
          description: Bulk not found
        "409":
          description: Bulk is not in a launchable state
  /benchmark/bulks/{bulkId}/retry-failed:
    post:
      operationId: retryBulkFailed
      tags: [benchmark]
      summary: Re-enqueue only failed/pending cells of this bulk's shard runs (FR-BLK-6; already-succeeded cells are never re-billed)
      parameters:
        - name: bulkId
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Retry started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bulk"
        "404":
          description: Bulk not found
        "409":
          description: Bulk is cancelled or still launching
  /benchmark/bulks/{bulkId}/cancel:
    post:
      operationId: cancelBulk
      tags: [benchmark]
      summary: Cancel a running bulk -- un-started cells are marked cancelled; in-flight provider requests complete and are recorded (FR-BLK-7)
      parameters:
        - name: bulkId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled bulk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bulk"
        "404":
          description: Bulk not found
        "409":
          description: Bulk is already in a terminal state
  /benchmark/bulks/{bulkId}/manifest:
    get:
      operationId: getBulkManifest
      tags: [benchmark]
      summary: Immutable bulk manifest -- the composition of every shard run's frozen manifest (FR-BLK-8, FR-REP1 replay evidence)
      parameters:
        - name: bulkId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bulk manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkManifest"
        "404":
          description: Bulk not found
  /benchmark/bulk-templates:
    get:
      operationId: listBulkTemplates
      tags: [benchmark]
      summary: List saved reusable bulk templates (FR-BLK-9)
      responses:
        "200":
          description: Bulk templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BulkTemplate"
    post:
      operationId: createBulkTemplate
      tags: [benchmark]
      summary: Save selection criteria as a named template. Criteria stay UNFROZEN -- a relative window like lastNDays is re-resolved at every launch (FR-BLK-9).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkTemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkTemplate"
        "409":
          description: Template name already in use
  /benchmark/bulk-templates/{templateId}/launch:
    post:
      operationId: launchBulkTemplate
      tags: [benchmark]
      summary: Launch a template -- re-resolves its relative date window against right now, then creates (and cost-gates) a fresh bulk (FR-BLK-9, AC-2.7)
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkTemplateLaunchInput"
      responses:
        "201":
          description: Bulk created from the template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bulk"
        "400":
          description: Nothing matched the resolved criteria
        "404":
          description: Template not found
...
components:
  schemas:
...
    Vertical:
      type: string
      enum: [rush, property_management, trucking]
...
    RunStatus:
      type: string
      enum: [queued, running, complete, blocked, failed, cancelled]
...
    BulkStatus:
      type: string
      # FR-BLK-4. partial = >=1 cell failed after retries; cancelled via
      # POST /benchmark/bulks/{id}/cancel.
      enum: [draft, estimating, awaiting_confirmation, running, complete, partial, failed, cancelled]
    BulkSelectionCriteria:
      type: object
      description: On a template this is unfrozen (lastNDays re-resolves per launch); on a bulk it is frozen at creation, with resolvedCallIds pinning the exact corpus slice (FR-BLK-1 vs FR-BLK-9).
      properties:
        vertical: { $ref: "#/components/schemas/Vertical" }
        assistantIds:
          type: array
          items: { type: string }
        accountLabel: { type: string }
        startedAtFrom: { type: string, format: date-time }
        startedAtTo: { type: string, format: date-time }
        lastNDays: { type: number, minimum: 1 }
        minDurationSeconds: { type: number, minimum: 0 }
        callIds:
          type: array
          items: { type: string }
        resolvedCallIds:
          type: array
          items: { type: string }
        resolvedAt: { type: ["string", "null"], format: date-time }
    Bulk:
      type: object
      required: [id, name, status, selectionCriteria, providerIds, shardSize, minDurationSeconds, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        status: { $ref: "#/components/schemas/BulkStatus" }
        selectionCriteria: { $ref: "#/components/schemas/BulkSelectionCriteria" }
        providerIds:
          type: array
          items: { type: string }
        shardSize: { type: number }
        minDurationSeconds: { type: number }
        estimatedCostCents: { type: ["number", "null"] }
        launchedByLabel: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
    BulkInput:
      type: object
      required: [criteria, providerIds]
      properties:
        # FR-BLK-2: defaults to the launch date (YYYY-MM-DD) when omitted.
        name: { type: string, minLength: 1 }
        criteria: { $ref: "#/components/schemas/BulkSelectionCriteria" }
        providerIds:
          type: array
          minItems: 1
          items: { type: string }
        shardSize: { type: number, minimum: 1, maximum: 500 }
        minDurationSeconds: { type: number, minimum: 0 }
        # FR-BLK-5: required true when the estimate exceeds the server cost
        # threshold; without it the bulk parks at awaiting_confirmation.
        confirm: { type: boolean }
    BulkProgress:
      type: object
      required: [callsTotal, callsRun, cellsTotal, cellsOk, cellsFailed, cellsPending, cellsCancelled, cellsSkippedPendingReview]
      properties:
        # FR-BLK-13 structured counts.
        callsTotal: { type: number }
        callsRun: { type: number }
        cellsTotal: { type: number }
        cellsOk: { type: number }
        cellsFailed: { type: number }
        cellsPending: { type: number }
        cellsCancelled: { type: number }
        cellsSkippedPendingReview: { type: number }
    BulkShardRun:
      type: object
      required: [id, shardIndex, status, callCount, createdAt]
      properties:
        id: { type: string }
        shardIndex: { type: number }
        status: { $ref: "#/components/schemas/RunStatus" }
        callCount: { type: number }
        createdAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
    BulkDetail:
      allOf:
        - $ref: "#/components/schemas/Bulk"
        - type: object
          required: [progress, runs]
          properties:
            progress: { $ref: "#/components/schemas/BulkProgress" }
            runs:
              type: array
              items: { $ref: "#/components/schemas/BulkShardRun" }
    BulkTemplate:
      type: object
      required: [id, name, selectionCriteria, providerIds, shardSize, minDurationSeconds, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        selectionCriteria: { $ref: "#/components/schemas/BulkSelectionCriteria" }
        providerIds:
          type: array
          items: { type: string }
        shardSize: { type: number }
        minDurationSeconds: { type: number }
        createdByLabel: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    BulkTemplateInput:
      type: object
      required: [name, criteria, providerIds]
      properties:
        name: { type: string, minLength: 1 }
        criteria: { $ref: "#/components/schemas/BulkSelectionCriteria" }
        providerIds:
          type: array
          minItems: 1
          items: { type: string }
        shardSize: { type: number, minimum: 1, maximum: 500 }
        minDurationSeconds: { type: number, minimum: 0 }
    BulkTemplateLaunchInput:
      type: object
      properties:
        # Override the auto-generated bulk name (default: template name + date).
        name: { type: string, minLength: 1 }
        confirm: { type: boolean }
    RunManifestCall:
      type: object
      required: [id, label, goldTranscriptSha256]
      properties:
        id: { type: string }
        label: { type: string }
        goldTranscriptSha256: { type: ["string", "null"] }
    RunManifestProvider:
      type: object
      required: [id, name, model, configSha256]
      properties:
        id: { type: string }
        name: { type: string }
        model: { type: string }
        configSha256: { type: string }
    RunManifest:
      type: object
      required: [manifestVersion, runId, scoringVersion, createdAt, calls, providers]
      properties:
        manifestVersion: { type: number }
        runId: { type: string }
        scoringVersion: { type: string }
        createdAt: { type: string, format: date-time }
        calls:
          type: array
          items: { $ref: "#/components/schemas/RunManifestCall" }
        providers:
          type: array
          items: { $ref: "#/components/schemas/RunManifestProvider" }
    BulkManifest:
      type: object
      required: [manifestVersion, bulkId, name, status, selectionCriteria, providerIds, shardSize, createdAt, runs]
      properties:
        manifestVersion: { type: number }
        bulkId: { type: string }
        name: { type: string }
        status: { $ref: "#/components/schemas/BulkStatus" }
        selectionCriteria: { $ref: "#/components/schemas/BulkSelectionCriteria" }
        providerIds:
          type: array
          items: { type: string }
        shardSize: { type: number }
        createdAt: { type: string, format: date-time }
        runs:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/RunManifest"
              - type: object
                properties:
                  shardIndex: { type: ["number", "null"] }
                  runStatus: { $ref: "#/components/schemas/RunStatus" }