# lib/api-zod/src/generated/types/bulkTemplate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API contract for the STT Benchmark Command Center
 * OpenAPI spec version: 0.1.0
 */
import type { BulkSelectionCriteria } from './bulkSelectionCriteria';

export interface BulkTemplate {
  id: string;
  name: string;
  selectionCriteria: BulkSelectionCriteria;
  providerIds: string[];
  shardSize: number;
  minDurationSeconds: number;
  /** @nullable */
  createdByLabel?: string | null;
  createdAt: Date;
  updatedAt?: Date;
}